refactor: 추가 카드 여부에서 y가 들어오는 경우 리팩토링
